+++ example/lib/main.dart
@@ -1,26 +1,13 @@
 // Created by Crt Vavros, copyright © 2022 ZeroPass. All rights reserved.
-// ignore_for_file: prefer_adjacent_string_concatenation, prefer_interpolation_to_compose_strings
+// MRTD Example App - Refactored with extracted widgets
 
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
-import 'dart:async';
-
-import 'package:dmrtd/dmrtd.dart';
-import 'package:dmrtd/extensions.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 import 'package:logging/logging.dart';
-import 'package:dmrtd/src/proto/can_key.dart';
-import 'package:intl/intl.dart';
+import 'package:dmrtd/extensions.dart';
 
-import 'package:mrtdeg/helpers/mrz_data.dart';
-import 'package:mrtdeg/view/scan_page.dart';
-import 'package:mrtdeg/widgets/mrtd_data_widget.dart';
-import 'package:mrtdeg/widgets/mrtd_access_widget.dart';
-import 'package:mrtdeg/widgets/dba_form.dart';
-import 'package:mrtdeg/widgets/pace_form.dart';
-import 'package:mrtdeg/models/mrtd_data.dart';
-import 'package:mrtdeg/utils/formatters.dart';
+import 'widgets/widgets.dart';
 
 void main() {
   Logger.root.level = Level.ALL;
@@ -36,674 +23,14 @@ class MrtdEgApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return PlatformApp(
-        localizationsDelegates: [
-          DefaultMaterialLocalizations.delegate,
-          DefaultCupertinoLocalizations.delegate,
-          DefaultWidgetsLocalizations.delegate,
-        ],
-        material: (_, __) => MaterialAppData(),
-        cupertino: (_, __) => CupertinoAppData(),
-        home: MrtdHomePage());
-  }
-}
-
-class MrtdHomePage extends StatefulWidget {
-  @override
-  // ignore: library_private_types_in_public_api
-  _MrtdHomePageState createState() => _MrtdHomePageState();
-}
-
-class _MrtdHomePageState extends State<MrtdHomePage>
-    with TickerProviderStateMixin {
-  var _alertMessage = "";
-  final _log = Logger("mrtdeg.app");
-  var _isNfcAvailable = false;
-  var _isReading = false;
-  final _mrzData = GlobalKey<FormState>();
-  final _canData = GlobalKey<FormState>();
-
-  // mrz data
-  final _docNumber = TextEditingController();
-  final _dob = TextEditingController(); // date of birth
-  final _doe = TextEditingController();
-  final _can = TextEditingController();
-  bool _checkBoxPACE = false;
-
-  MrtdData? _mrtdData;
-
-  final NfcProvider _nfc = NfcProvider();
-
-  // ignore: unused_field
-  late Timer _timerStateUpdater;
-  final _scrollController = ScrollController();
-  late final TabController _tabController;
-
-  @override
-  void initState() {
-    super.initState();
-
-    _tabController = TabController(length: 2, vsync: this);
-    //_tabController.addListener(_handleTabSelection);
-
-    SystemChrome.setPreferredOrientations([
-      DeviceOrientation.portraitUp,
-      DeviceOrientation.portraitDown,
-    ]);
-
-    _initPlatformState();
-
-    // Update platform state every 3 sec
-    _timerStateUpdater = Timer.periodic(Duration(seconds: 3), (Timer t) {
-      _initPlatformState();
-    });
-  }
-
-  // Platform messages are asynchronous, so we initialize in an async method.
-  Future<void> _initPlatformState() async {
-    bool isNfcAvailable;
-    try {
-      NfcStatus status = await NfcProvider.nfcStatus;
-      isNfcAvailable = status == NfcStatus.enabled;
-    } on PlatformException {
-      isNfcAvailable = false;
-    }
-
-    // If the widget was removed from the tree while the asynchronous platform
-    // message was in flight, we want to discard the reply rather than calling
-    // setState to update our non-existent appearance.
-    if (!mounted) return;
-
-    setState(() {
-      _isNfcAvailable = isNfcAvailable;
-    });
-  }
-
-  DateTime? _getDOBDate() {
-    if (_dob.text.isEmpty) {
-      return null;
-    }
-    return DateFormat.yMd().parse(_dob.text);
-  }
-
-  DateTime? _getDOEDate() {
-    if (_doe.text.isEmpty) {
-      return null;
-    }
-    return DateFormat.yMd().parse(_doe.text);
-  }
-
-  Future<String?> _pickDate(BuildContext context, DateTime firstDate,
-      DateTime initDate, DateTime lastDate) async {
-    final locale = Localizations.localeOf(context);
-    final DateTime? picked = await showDatePicker(
-        context: context,
-        firstDate: firstDate,
-        initialDate: initDate,
-        lastDate: lastDate,
-        locale: locale);
-
-    if (picked != null) {
-      return DateFormat.yMd().format(picked);
-    }
-    return null;
-  }
-
-  void _readMRZPressed() async {
-    final MRZResult? mrzResult = await showDialog(context: context, builder: (context) {
-      return Material(child: ScannerPage());
-    });
-    if (mrzResult != null) {
-      setState(() {
-      _docNumber.text = mrzResult.documentNumber;
-      _dob.text = DateFormat.yMd().format(mrzResult.birthDate);
-      _doe.text = DateFormat.yMd().format(mrzResult.expiryDate);
-      });
-    }
-  }
-
-  void _buttonPressed() async {
-    print("Button pressed");
-    //Check on what tab we are
-    if (_tabController.index == 0) {
-      //DBA tab
-      String errorText = "";
-      if (_doe.text.isEmpty) {
-        errorText += "Please enter date of expiry!\n";
-      }
-      if (_dob.text.isEmpty) {
-        errorText += "Please enter date of birth!\n";
-      }
-      if (_docNumber.text.isEmpty) {
-        errorText += "Please enter passport number!";
-      }
-
-      setState(() {
-        _alertMessage = errorText;
-      });
-      //If there is an error, just jump out of the function
-      if (errorText.isNotEmpty) return;
-
-      final bacKeySeed = DBAKey(_docNumber.text, _getDOBDate()!, _getDOEDate()!,
-          paceMode: _checkBoxPACE);
-      _readMRTD(accessKey: bacKeySeed, isPace: _checkBoxPACE);
-    } else {
-      //PACE tab
-      String errorText = "";
-      if (_can.text.isEmpty) {
-        errorText = "Please enter CAN number!";
-      } else if (_can.text.length != 6) {
-        errorText = "CAN number must be exactly 6 digits long!";
-      }
-
-      setState(() {
-        _alertMessage = errorText;
-      });
-      //If there is an error, just jump out of the function
-      if (errorText.isNotEmpty) return;
-
-      final canKeySeed = CanKey(_can.text);
-      _readMRTD(accessKey: canKeySeed, isPace: true);
-    }
-  }
-
-  void _readMRTD({required AccessKey accessKey, bool isPace = false}) async {
-    try {
-      setState(() {
-        _mrtdData = null;
-        _alertMessage = "Waiting for Passport tag ...";
-        _isReading = true;
-      });
-      try {
-        bool demo = false;
-        if (!demo) {
-          await _nfc.connect(
-              iosAlertMessage: "Hold your phone near Biometric Passport");
-        }
-
-        final passport = Passport(_nfc);
-
-        setState(() {
-          _alertMessage = "Reading Passport ...";
-        });
-
-        _nfc.setIosAlertMessage("Trying to read EF.CardAccess ...");
-        final mrtdData = MrtdData();
-
-        try {
-          mrtdData.cardAccess = await passport.readEfCardAccess();
-        } on PassportError {
-          //if (e.code != StatusWord.fileNotFound) rethrow;
-        }
-
-        _nfc.setIosAlertMessage("Trying to read EF.CardSecurity ...");
-
-        try {
-          //mrtdData.cardSecurity = await passport.readEfCardSecurity();
-        } on PassportError {
-          //if (e.code != StatusWord.fileNotFound) rethrow;
-        }
-
-        _nfc.setIosAlertMessage("Initiating session with PACE...");
-        //set MrtdData
-        mrtdData.isPACE = isPace;
-        mrtdData.isDBA = accessKey.PACE_REF_KEY_TAG == 0x01 ? true : false;
-
-        if (isPace) {
-          //PACE session
-          await passport.startSessionPACE(accessKey, mrtdData.cardAccess!);
-        } else {
-          //BAC session
-          await passport.startSession(accessKey as DBAKey);
-        }
-
-        _nfc.setIosAlertMessage(formatProgressMsg("Reading EF.COM ...", 0));
-        mrtdData.com = await passport.readEfCOM();
-
-        _nfc.setIosAlertMessage(
-            formatProgressMsg("Reading Data Groups ...", 20));
-
-        if (mrtdData.com!.dgTags.contains(EfDG1.TAG)) {
-          mrtdData.dg1 = await passport.readEfDG1();
-        }
-
-        if (mrtdData.com!.dgTags.contains(EfDG2.TAG)) {
-          mrtdData.dg2 = await passport.readEfDG2();
-        }
-
-        // To read DG3 and DG4 session has to be established with CVCA certificate (not supported).
-        // if(mrtdData.com!.dgTags.contains(EfDG3.TAG)) {
-        //   mrtdData.dg3 = await passport.readEfDG3();
-        // }
-
-        // if(mrtdData.com!.dgTags.contains(EfDG4.TAG)) {
-        //   mrtdData.dg4 = await passport.readEfDG4();
-        // }
-
-        if (mrtdData.com!.dgTags.contains(EfDG5.TAG)) {
-          mrtdData.dg5 = await passport.readEfDG5();
-        }
-
-        if (mrtdData.com!.dgTags.contains(EfDG6.TAG)) {
-          mrtdData.dg6 = await passport.readEfDG6();
-        }
-
-        if (mrtdData.com!.dgTags.contains(EfDG7.TAG)) {
-          mrtdData.dg7 = await passport.readEfDG7();
-        }
-
-        if (mrtdData.com!.dgTags.contains(EfDG8.TAG)) {
-          mrtdData.dg8 = await passport.readEfDG8();
-        }
-
-        if (mrtdData.com!.dgTags.contains(EfDG9.TAG)) {
-          mrtdData.dg9 = await passport.readEfDG9();
-        }
-
-        if (mrtdData.com!.dgTags.contains(EfDG10.TAG)) {
-          mrtdData.dg10 = await passport.readEfDG10();
-        }
-
-        if (mrtdData.com!.dgTags.contains(EfDG11.TAG)) {
-          mrtdData.dg11 = await passport.readEfDG11();
-        }
-
-        if (mrtdData.com!.dgTags.contains(EfDG12.TAG)) {
-          mrtdData.dg12 = await passport.readEfDG12();
-        }
-
-        if (mrtdData.com!.dgTags.contains(EfDG13.TAG)) {
-          mrtdData.dg13 = await passport.readEfDG13();
-        }
-
-        if (mrtdData.com!.dgTags.contains(EfDG14.TAG)) {
-          mrtdData.dg14 = await passport.readEfDG14();
-        }
-
-        if (mrtdData.com!.dgTags.contains(EfDG15.TAG)) {
-          mrtdData.dg15 = await passport.readEfDG15();
-          _nfc.setIosAlertMessage(formatProgressMsg("Doing AA ...", 60));
-          mrtdData.aaSig = await passport.activeAuthenticate(Uint8List(8));
-        }
-
-        if (mrtdData.com!.dgTags.contains(EfDG16.TAG)) {
-          mrtdData.dg16 = await passport.readEfDG16();
-        }
-
-        _nfc.setIosAlertMessage(formatProgressMsg("Reading EF.SOD ...", 80));
-        mrtdData.sod = await passport.readEfSOD();
-
-        setState(() {
-          _mrtdData = mrtdData;
-        });
-
-        setState(() {
-          _alertMessage = "";
-        });
-
-        _scrollController.animateTo(300.0,
-            duration: Duration(milliseconds: 500), curve: Curves.ease);
-      } on Exception catch (e) {
-        final se = e.toString().toLowerCase();
-        String alertMsg = "An error has occurred while reading Passport!";
-        if (e is PassportError) {
-          if (se.contains("security status not satisfied")) {
-            alertMsg =
-                "Failed to initiate session with passport.\nCheck input data!";
-          }
-          _log.error("PassportError: ${e.message}");
-        } else {
-          _log.error(
-              "An exception was encountered while trying to read Passport: $e");
-        }
-
-        if (se.contains('timeout')) {
-          alertMsg = "Timeout while waiting for Passport tag";
-        } else if (se.contains("tag was lost")) {
-          alertMsg = "Tag was lost. Please try again!";
-        } else if (se.contains("invalidated by user")) {
-          alertMsg = "";
-        }
-
-        setState(() {
-          _alertMessage = alertMsg;
-        });
-      } finally {
-        if (_alertMessage.isNotEmpty) {
-          await _nfc.disconnect(iosErrorMessage: _alertMessage);
-        } else {
-          await _nfc.disconnect(
-              iosAlertMessage: formatProgressMsg("Finished", 100));
-        }
-        setState(() {
-          _isReading = false;
-        });
-      }
-    } on Exception catch (e) {
-      _log.error("Read MRTD error: $e");
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return PlatformProvider(
-        builder: (BuildContext context) => _buildPage(context));
-  }
-
-  bool _disabledInput() {
-    //return true;
-    return _isReading || !_isNfcAvailable;
-  }
-
-
-  List<Widget> _mrtdDataWidgets() {
-    List<Widget> list = [];
-    if (_mrtdData == null) return list;
-
-    if (_mrtdData!.isPACE != null && _mrtdData!.isDBA != null) {
-      list.add(MrtdAccessWidget(
-          header: "Access protocol",
-          collapsedText: '',
-          isDBA: _mrtdData!.isDBA!,
-          isPACE: _mrtdData!.isPACE!));
-    }
-
-    if (_mrtdData!.cardAccess != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.CardAccess',
-          collapsedText: '',
-          dataText: _mrtdData!.cardAccess!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.cardSecurity != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.CardSecurity',
-          collapsedText: '',
-          dataText: _mrtdData!.cardSecurity!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.sod != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.SOD',
-          collapsedText: '',
-          dataText: _mrtdData!.sod!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.com != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.COM',
-          collapsedText: '',
-          dataText: formatEfCom(_mrtdData!.com!)));
-    }
-
-    if (_mrtdData!.dg1 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG1',
-          collapsedText: '',
-          dataText: formatMRZ(_mrtdData!.dg1!.mrz)));
-    }
-
-    if (_mrtdData!.dg2 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG2',
-          collapsedText: '',
-          dataText: _mrtdData!.dg2!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.dg3 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG3',
-          collapsedText: '',
-          dataText: _mrtdData!.dg3!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.dg4 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG4',
-          collapsedText: '',
-          dataText: _mrtdData!.dg4!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.dg5 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG5',
-          collapsedText: '',
-          dataText: _mrtdData!.dg5!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.dg6 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG6',
-          collapsedText: '',
-          dataText: _mrtdData!.dg6!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.dg7 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG7',
-          collapsedText: '',
-          dataText: _mrtdData!.dg7!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.dg8 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG8',
-          collapsedText: '',
-          dataText: _mrtdData!.dg8!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.dg9 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG9',
-          collapsedText: '',
-          dataText: _mrtdData!.dg9!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.dg10 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG10',
-          collapsedText: '',
-          dataText: _mrtdData!.dg10!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.dg11 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG11',
-          collapsedText: '',
-          dataText: _mrtdData!.dg11!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.dg12 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG12',
-          collapsedText: '',
-          dataText: _mrtdData!.dg12!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.dg13 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG13',
-          collapsedText: '',
-          dataText: _mrtdData!.dg13!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.dg14 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG14',
-          collapsedText: '',
-          dataText: _mrtdData!.dg14!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.dg15 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG15',
-          collapsedText: '',
-          dataText: _mrtdData!.dg15!.toBytes().hex()));
-    }
-
-    if (_mrtdData!.aaSig != null) {
-      list.add(MrtdDataWidget(
-          header: 'Active Authentication signature',
-          collapsedText: '',
-          dataText: _mrtdData!.aaSig!.hex()));
-    }
-
-    if (_mrtdData!.dg16 != null) {
-      list.add(MrtdDataWidget(
-          header: 'EF.DG16',
-          collapsedText: '',
-          dataText: _mrtdData!.dg16!.toBytes().hex()));
-    }
-
-    return list;
-  }
-
-  PlatformScaffold _buildPage(BuildContext context) => PlatformScaffold(
-        appBar: PlatformAppBar(title: Text('MRTD Example App')),
-        iosContentPadding: false,
-        iosContentBottomPadding: false,
-        body: Material(
-          child: SafeArea(
-            child: Padding(
-              padding: EdgeInsets.all(8.0),
-              child: SingleChildScrollView(
-                controller: _scrollController,
-                child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.stretch,
-                    children: <Widget>[
-                      _buildForm(context),
-                      SizedBox(height: 20),
-                      PlatformElevatedButton(
-                        // btn Read MRTD
-                        onPressed: _buttonPressed,
-                        child: PlatformText(
-                            _isReading ? 'Reading ...' : 'Read Passport'),
-                      ),
-                      SizedBox(height: 20),
-                      PlatformElevatedButton(
-                        // btn Read MRTD
-                        onPressed: _readMRZPressed,
-                        child: PlatformText(
-                            _isReading ? 'Reading ...' : 'Scan MRZ'),
-                      ),
-                      SizedBox(height: 20),
-                      Row(children: <Widget>[
-                        Text('NFC available:',
-                            style: TextStyle(
-                                fontSize: 18.0, fontWeight: FontWeight.bold)),
-                        SizedBox(width: 4),
-                        Text(_isNfcAvailable ? "Yes" : "No",
-                            style: TextStyle(fontSize: 18.0))
-                      ]),
-                      SizedBox(height: 15),
-                      Text(_alertMessage,
-                          textAlign: TextAlign.center,
-                          style: TextStyle(
-                              fontSize: 15.0, fontWeight: FontWeight.bold)),
-                      SizedBox(height: 15),
-                      Padding(
-                        padding: EdgeInsets.all(8.0),
-                        child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: <Widget>[
-                              Text(_mrtdData != null ? "Passport Data:" : "",
-                                  textAlign: TextAlign.center,
-                                  style: TextStyle(
-                                      fontSize: 15.0,
-                                      fontWeight: FontWeight.bold)),
-                              Padding(
-                                  padding: EdgeInsets.only(
-                                      left: 16.0, top: 8.0, bottom: 8.0),
-                                  child: Column(
-                                      crossAxisAlignment:
-                                          CrossAxisAlignment.start,
-                                      children: _mrtdDataWidgets()))
-                            ]),
-                      ),
-                    ]),
-              ),
-            ),
-          ),
-        ),
-      );
-
-  Widget _buildForm(BuildContext context) {
-    return Column(
-      children: [
-        TabBar(
-          controller: _tabController,
-          labelColor: Colors.blue,
-          tabs: const [
-            Tab(text: 'DBA'),
-            Tab(text: 'PACE'),
-          ],
-        ),
-        Container(
-          height: 350,
-          child: TabBarView(
-            controller: _tabController,
-            children: [
-              Card(
-                borderOnForeground: false,
-                elevation: 0,
-                color: Colors.white,
-                margin: const EdgeInsets.all(16.0),
-                child: Form(
-                  key: _mrzData,
-                  child: DbaForm(
-                    docNumberController: _docNumber,
-                    dobController: _dob,
-                    doeController: _doe,
-                    paceChecked: _checkBoxPACE,
-                    disabled: _disabledInput(),
-                    onPaceChanged: (val) {
-                      setState(() => _checkBoxPACE = !_checkBoxPACE);
-                    },
-                    onPickDob: () async {
-                      FocusScope.of(context).requestFocus(FocusNode());
-                      final now = DateTime.now();
-                      final firstDate =
-                          DateTime(now.year - 90, now.month, now.day);
-                      final lastDate =
-                          DateTime(now.year - 15, now.month, now.day);
-                      final initDate = _getDOBDate();
-                      final date = await _pickDate(
-                          context, firstDate, initDate ?? lastDate, lastDate);
-                      FocusScope.of(context).requestFocus(FocusNode());
-                      return date;
-                    },
-                    onPickDoe: () async {
-                      FocusScope.of(context).requestFocus(FocusNode());
-                      final now = DateTime.now();
-                      final firstDate =
-                          DateTime(now.year, now.month, now.day + 1);
-                      final lastDate =
-                          DateTime(now.year + 10, now.month + 6, now.day);
-                      final initDate = _getDOEDate();
-                      final date = await _pickDate(
-                          context, firstDate, initDate ?? firstDate, lastDate);
-                      FocusScope.of(context).requestFocus(FocusNode());
-                      return date;
-                    },
-                  ),
-                ),
-              ),
-              Card(
-                borderOnForeground: false,
-                elevation: 0,
-                color: Colors.white,
-                margin: const EdgeInsets.all(16.0),
-                child: Form(
-                  key: _canData,
-                  child: PaceForm(
-                    canController: _can,
-                    disabled: _disabledInput(),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
+      localizationsDelegates: const [
+        DefaultMaterialLocalizations.delegate,
+        DefaultCupertinoLocalizations.delegate,
+        DefaultWidgetsLocalizations.delegate,
       ],
+      material: (_, __) => MaterialAppData(),
+      cupertino: (_, __) => CupertinoAppData(),
+      home: MrtdHomePage(),
     );
   }
 }
