+++ example/lib/widgets/displays/mrtd_data_widget.dart
@@ -1,19 +1,23 @@
+// Created by Crt Vavros, copyright © 2022 ZeroPass. All rights reserved.
+// MRTD data display widget with expandable panels
+
 import 'package:expandable/expandable.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 
+/// Widget to display MRTD data in expandable panels
 class MrtdDataWidget extends StatelessWidget {
   final String header;
   final String collapsedText;
   final String dataText;
 
   const MrtdDataWidget({
-    super.key,
+    Key? key,
     required this.header,
     required this.collapsedText,
     required this.dataText,
-  });
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -38,11 +42,14 @@ class MrtdDataWidget extends StatelessWidget {
           crossAxisAlignment: CrossAxisAlignment.stretch,
           children: [
             PlatformTextButton(
-              padding: const EdgeInsets.all(8),
-              onPressed: () => Clipboard.setData(ClipboardData(text: dataText)),
               child: const Text('Copy'),
+              onPressed: () => Clipboard.setData(ClipboardData(text: dataText)),
+              padding: const EdgeInsets.all(8),
             ),
-            SelectableText(dataText, textAlign: TextAlign.left),
+            SelectableText(
+              dataText,
+              textAlign: TextAlign.left,
+            ),
           ],
         ),
       ),
