+++ example/lib/widgets/widgets.dart
@@ -1,0 +1,10 @@
+// Main widgets export file
+export 'common/alert_message_widget.dart';
+export 'common/nfc_status_widget.dart';
+export 'displays/access_protocol_widget.dart';
+export 'displays/mrtd_data_list_widget.dart';
+export 'displays/mrtd_data_widget.dart';
+export 'forms/auth_form_widget.dart';
+export 'forms/can_form_widget.dart';
+export 'forms/dba_form_widget.dart';
+export 'pages/mrtd_home_page.dart';
