+++ pubspec.yaml
@@ -21,6 +21,7 @@ dependencies:
   expandable: ^5.0.1
   fixnum: ^1.1.1
   flutter_nfc_kit: ^3.6.0-rc.6
+  intl: ^0.20.2
     #git:
   #  url: https://github.com/ZeroPass/flutter_nfc_kit.git
   logging: ^1.3.0
